+++ CHANGELOG.md
@@ -7,7 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
-For a full diff see [`0.20.0...main`][0.20.0...main].
+For a full diff see [`0.21.0...main`][0.21.0...main].
+
+## [`0.21.0`][0.21.0] - 2026-06-13
+
+For a full diff see [`0.20.0...0.21.0`][0.20.0...0.21.0].
 
 ### Added
 
@@ -397,6 +401,7 @@ For a full diff see [`0.2.2...0.3.0`][0.2.2...0.3.0].
 [0.19.0]: https://github.com/weph/obsidian-bookshelf/releases/tag/0.19.0
 [0.19.1]: https://github.com/weph/obsidian-bookshelf/releases/tag/0.19.1
 [0.20.0]: https://github.com/weph/obsidian-bookshelf/releases/tag/0.20.0
+[0.21.0]: https://github.com/weph/obsidian-bookshelf/releases/tag/0.21.0
 [0.2.2...0.3.0]: https://github.com/weph/obsidian-bookshelf/compare/0.2.2...0.3.0
 [0.3.0...0.3.1]: https://github.com/weph/obsidian-bookshelf/compare/0.3.0...0.3.1
 [0.3.1...0.4.0]: https://github.com/weph/obsidian-bookshelf/compare/0.3.1...0.4.0
@@ -420,4 +425,5 @@ For a full diff see [`0.2.2...0.3.0`][0.2.2...0.3.0].
 [0.18.0...0.19.0]: https://github.com/weph/obsidian-bookshelf/compare/0.18.0...0.19.0
 [0.19.0...0.19.1]: https://github.com/weph/obsidian-bookshelf/compare/0.19.0...0.19.1
 [0.19.1...0.20.0]: https://github.com/weph/obsidian-bookshelf/compare/0.19.1...0.20.0
-[0.20.0...main]: https://github.com/weph/obsidian-bookshelf/compare/0.20.0...main
+[0.20.0...0.21.0]: https://github.com/weph/obsidian-bookshelf/compare/0.20.0...0.21.0
+[0.21.0...main]: https://github.com/weph/obsidian-bookshelf/compare/0.21.0...main